backend/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.38.0
textblob>=0.19.0
supabase>=2.27.2
python-jose[cryptography]>=3.5.0
openai>=2.14.0
httpx>=0.28.1
python-dotenv>=1.2.1
pydantic>=2.12.0
